Add federation client and server methods for bulk user info
We can't just use a federation query here as we need to send a JSON body of
information. This makes things slightly more complex, but it basically boils
down to creating server and client methods.

The server method had some slight annoyance in that you're supposed to do all
the logic in FederationServer, however the logic already exists in the
data_store. You can only access the data_store from the handler, so... cue
relatively useless method in FederationServer.