Add room freeze functionality when the last admin leaves
When a membership event is checked by the room access rules, we check if it's both a
leave membership, as well as whether this belongs to the last admin (PL>=100) user in
a room. If it does, we freeze the room.

When freezing the room, we take the existing power level state event of the room, modify
it so that everything requires PL100 (meaning the admin could come back and fix the room
if desired), sends the state event, and then actually leaves.

We are careful to only do this if the user leaving is a local user. Otherwise, we wouldn't
be able to puppet them and send a state event into the room. To enable this, we also now
pass server_name to room access rules.