Fix new power level event being deduplicated, thus not persisted
Turns out we were overridding the content of the old power level event in the room, which
means that when it came time to persist the new power level event, the code compared the
content of the two and determined that the new state event was a duplicated, and thus it
wasn't persisted - so clients didn't see it down sync, and the new power levels were not
enforced after a server restart.

Copying the old power levels dict instead of using it directly solves the issue.