Switch account_validity templates to read_template method
We're doing this as we're planning to inject account expiry dates into the templates
(so that they can show "Your account is valid until XX/YY/ZZ"). Since this is
only information we'll have during server operation, we need to switch the config
variables to be templates instead.

Also nice as we no longer need to hardcoded the template content inline.

The public_baseurl change was due to self.public_baseurl raising an
AssertionError if we try to call it before it exists, which happens to
be the case for the account validity config code. Note that we override
__getattr__ above.