Return the new template if the user is trying to reuse a token.
We check if the user is re-using a token by (ab)using the 'email_sent' column associated with
each user's account validity information. When a token is first created and emailed to a user,
'email_sent' is set to True. When a user comes back to us with this token, we renew the account
and set 'email_sent' to False - but critically we keep the token around.

If the user comes back again with the same token, we'll see that although the token is valid,
'email_sent' is False. From this we can learn that the token was already used to renew the
account. We call this case a stale token, and return a new template to the user. The new
template simply informs the user when their account will expire.

This is much nicer UX than the 'invalid token!' users would get when they tried a token
twice.