Modify reg tests to check re-using an existing account validity token.
Note that the variable for this test include an account expiring after 1 week of it
last being renewed (or created), and sending the user a renewal token 2 days before
the account is set to expire.

The change here is that after we successfully renew the account, move forward and
try to renew it again. We should get a success, and the new template returned,
but the account's expiration date should not have been extended.