Move account validity template reading to RegistrationConfig
Calling `self.read_templates` from AccountValidityConfig does not work as
expected as it does not have state such as self.public_baseurl properly set up.
This is due to AccountValidityConfig not being initialised like other config
classes, but instead simply in the body of RegistrationConfig.read_config.

I think there's an argument for initialising it properly, instead of as
a part of registration config. Perhaps even putting it in its own file.
But that should be left to a separate PR.