Use hs.get_simple_http_client instead of creating a new SimpleHttpClient instance
This reverts a small change made in
https://github.com/matrix-org/synapse/commit/e08ea43463bacd5efacbf6c790c6be0f3cd06ce6
that switches from using the HS's SimpleHttpClient to creating a new one. There isn't
a point in doing so without providing any arguments, and doing so breaks the way a
test checks that a call was made, so let's change it back.

I plan to port this change back to mainline as well.